Changes int size to Long and full Spring Router statistic

+++ src/main/java/com/citycon/statistic/repositories/CityStatisticRepository.java
@@ -20,14 +20,14 @@ public class CityStatisticRepository extends AbstractRepository {
         this.namedParameterDao = namedParameterDao;
     }
 
-    public int getCount() {
+    public Long getCount() {
         return super.getCount(TABLE_NAME);
     }
 
-    public int getCountriesCount() {
+    public Long getCountriesCount() {
         String query = "SELECT COUNT(DISTINCT Country) FROM "+TABLE_NAME;
 
-        int count = dao.queryForObject(query, Integer.class);
+        Long count = dao.queryForObject(query, Long.class);
         return count;
     }
